+++ Install/basic_environment.yaml
@@ -22,16 +22,17 @@ dependencies:
   - ninja
   - openblas-devel
   - yacs==0.1.8
-  - numpy==1.24.3
   - networkx==3.1
   - rtree==1.2.0
   - scipy==1.10.1
   - matplotlib
 
   - pip:
+      - numpy==1.23.5
       - opencv-python
       - pybullet==3.2.6
       - open3d
+      - graspnetAPI
       - ompl-1.6.0-cp38-cp38-manylinux_2_28_x86_64.whl
 
       - -i https://pypi.tuna.tsinghua.edu.cn/simple
