+++ Install/pythonpath.sh
@@ -1,11 +1,17 @@
 #!/bin/bash
 
 # Add this project to the python search path
-script_dir=$(dirname "$(readlink -f "$0")")
-project_dir=$(dirname "$script_dir")
+Install_dir=$(dirname "$(readlink -f "$0")")
+project_dir=$(dirname "$Install_dir")
 if [ -z "$PYTHONPATH" ] || ! echo "$PYTHONPATH" | grep -q "$project_dir"; then
-    # If PYTHONPATH doesn't exist, set it to project directory
-    echo "export PYTHONPATH=\$PYTHONPATH:$project_dir" >> ~/.bashrc
+    PYTHONPATH="\$PYTHONPATH:$project_dir"
 fi
-# Create conda environment
-mamba env create -f basic_environment.yaml
+
+# Add AnyGrasp .so files to the python search path
+Anygrasp_dir="$project_dir/Motion_Planning/Manipulation/Grasp/Grasp_Pose_Estimation/AnyGrasp"
+if ! echo "$PYTHONPATH" | grep -q "$Anygrasp_dir"; then
+    PYTHONPATH="$PYTHONPATH:$Anygrasp_dir"
+fi
+
+echo "export PYTHONPATH=$PYTHONPATH" >> ~/.bashrc
+
