+++ Motion_Planning/Manipulation/Grasp/Grasp_Pose_Estimation/AnyGrasp/src/license/KuiYang.public_key
@@ -1,0 +1,9 @@
+-----BEGIN PUBLIC KEY-----
+MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAw097GUzgmxPa50UaA0Ua
+W8ur1WtHkDK2JwI7VTapxihHvnRT3WXWZEMbGrETi+OxUmSuj5ja3IgEAfdsvqZ+
+Q1n/SUwihXmHKoxbOoaF/YVJIRKOH7NyXPz6xq8QFKj5tqfMAn8+LH7ph69cWyFs
+/Lz85AS+nAifakDeSSwTsmh921P5me4uONPoca9pWBBunFa5sbHrbBM6NHG9g+Si
+S8jsaJtlpt7Rh4S1dnmHWkaStccBjITi61zO68IldI69I+ZYZOn/z/G5JukpZfzH
+Jb+CCxIYOMio4AeQQ8hAHhmnv2tRgdpeQDt3vLDkBd/Hdx1g3D48a5hU8ZEUMO2k
+MQIDAQAB
+-----END PUBLIC KEY-----
