+++ Motion_Planning/Manipulation/Grasp/Grasp_Pose_Estimation/AnyGrasp/src/license/licenseCfg.json
@@ -1,0 +1,14 @@
+{
+    "version": "3.1.0",
+    "feature_id": "18324784934840148220",
+    "public_key": "KuiYang.public_key",
+    "signature": "KuiYang.signature",
+    "license": "KuiYang.lic",
+    "toolbox": [
+        {
+            "name": "Basic",
+            "type": "PERMANENT",
+            "data": ""
+        }
+    ]
+}
