+++ Motion_Planning/Manipulation/Grasp/Grasp_Pose_Estimation/AnyGrasp/src/test_data/test_depth.png
