+++ Motion_Planning/Manipulation/Grasp/Grasp_Pose_Estimation/AnyGrasp/src/utils.py
@@ -1,4 +1,5 @@
 import numpy as np
+import open3d as o3d
 from PIL import ImageDraw
 
 def sample_points(points, sampling_rate=1):
