--- Motion_Planning/Manipulation/Grasp/Grasp_Pose_Estimation/AnyGrasp/utils/types.py
@@ -1,4 +1,0 @@
-from typing import Union, Tuple
-
-Number = Union[int, float]
-Bbox = Tuple[int, int, int, int]
