+++ Motion_Planning/Manipulation/Grasp/Grasp_Pose_Estimation/__init__.py
@@ -1,2 +1,1 @@
-from .AnyGrasp.AnyGrasp import AnyGrasp
-from .AnyGrasp.utils.camera import CameraParameters
+from .AnyGrasp.src.AnyGrasp import AnyGrasp
