+++ Perception/Object_detection/__init__.py
@@ -1,2 +1,1 @@
-from Lang_SAM.Lang_SAM import Lang_SAM
-from .object_detection import object_detection
+from .Lang_SAM.Lang_SAM import Lang_SAM
